move products into interface/src
