+++ RecoLocalCalo/EcalRecProducers/BuildFile.xml
@@ -2,6 +2,11 @@
 <use   name="FWCore/Framework"/>
 <use   name="clhep"/>
 <use   name="DataFormats/EcalRecHitSoA"/>
+<use   name="CondFormats/EcalObjects"/>
+<use   name="HeterogeneousCore/CUDACore"/>
+<use   name="HeterogeneousCore/CUDAUtilities"/>
+<use   name="cuda-api-wrappers"/>
+<use   name="cuda"/>
 <export>
   <lib   name="1"/>
 </export>
