+++ RecoLocalCalo/EcalRecProducers/interface/EcalGainRatiosGPU.h
