+++ RecoLocalCalo/EcalRecProducers/interface/EcalPedestalsGPU.h
