+++ RecoLocalCalo/EcalRecProducers/interface/EcalPulseCovariancesGPU.h
