+++ RecoLocalCalo/EcalRecProducers/interface/EcalPulseShapesGPU.h
