+++ RecoLocalCalo/EcalRecProducers/interface/EcalSamplesCorrelationGPU.h
