+++ RecoLocalCalo/EcalRecProducers/plugins/EcalESProducerGPU.h
@@ -17,9 +17,9 @@ public:
     explicit EcalESProducerGPU(edm::ParameterSet const& ps) 
         : label_{ps.getParameter<std::string>("label")}
     {
+        std::cout << __PRETTY_FUNCTION__ << std::endl;
         std::string name = ps.getParameter<std::string>("ComponentName");
-        setWhatProduced(this, &EcalESProducerGPU<Target, Source, Record>::produce,
-                        edm::es::Label{name});
+        setWhatProduced(this, name);
     }
    
     std::unique_ptr<Target> produce(Record const& record) {
