+++ RecoLocalCalo/EcalRecProducers/src/EcalGainRatiosGPU.cc
@@ -1,4 +1,4 @@
-#include "EcalGainRatiosGPU.h"
+#include "RecoLocalCalo/EcalRecProducers/interface/EcalGainRatiosGPU.h"
 
 #include "FWCore/Utilities/interface/typelookup.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/cudaCheck.h"
