+++ RecoLocalCalo/EcalRecProducers/src/EcalSamplesCorrelationGPU.cc
@@ -1,4 +1,4 @@
-#include "EcalSamplesCorrelationGPU.h"
+#include "RecoLocalCalo/EcalRecProducers/interface/EcalSamplesCorrelationGPU.h"
 
 #include "FWCore/Utilities/interface/typelookup.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/cudaCheck.h"
